+++ tests/test_ac1015/test_basic_graphics.py
@@ -18,6 +18,17 @@ class SetupDrawing(unittest.TestCase):
         self.layout = self.dwg.modelspace()
 
 
+class TestGraphicsDefaultSettings(SetupDrawing):
+    def test_default_settings(self):
+        line = self.layout.add_line((0, 0), (1, 1))
+        self.assertEqual('0', line.dxf.layer)
+        self.assertEqual(256, line.dxf.color)
+        self.assertEqual('BYLAYER', line.dxf.linetype)
+        self.assertEqual(1.0, line.dxf.ltscale)
+        self.assertEqual(0, line.dxf.invisible)
+        self.assertEqual((0.0, 0.0, 1.0), line.dxf.extrusion)
+
+
 class TestBasicEntities(SetupDrawing):
     def test_iter_layout(self):
         self.layout.add_line((0, 0), (1, 1))
@@ -56,15 +67,16 @@ class TestBasicEntities(SetupDrawing):
         self.assertEqual((1, 1), trace[3])
 
     def test_create_3dface(self):
-        trace = self.layout.add_3Dface([(0, 0), (1, 0), (1, 1), (0, 1)])
-        self.assertEqual((0, 0), trace.dxf.vtx0)
-        self.assertEqual((1, 0), trace[1])
-        self.assertEqual((1, 1), trace.dxf.vtx2)
-        self.assertEqual((0, 1), trace[3])
+        trace = self.layout.add_3Dface([(0, 0, 0), (1, 0, 0), (1, 1, 0), (0, 1, 0)])
+        self.assertEqual((0, 0, 0), trace.dxf.vtx0)
+        self.assertEqual((1, 0, 0), trace[1])
+        self.assertEqual((1, 1, 0), trace.dxf.vtx2)
+        self.assertEqual((0, 1, 0), trace[3])
 
     def test_create_text(self):
         text = self.layout.add_text('text')
         self.assertEqual('text', text.dxf.text)
+
 
 class TestText(SetupDrawing):
     def test_create_text(self):